convert duration matrix from float to int

+++ DARP_IPS/src/dynamicMain.cpp
@@ -11,7 +11,7 @@
 
 
 using namespace std::chrono;
-vector2D<float> durationMatrix_;
+vector2D<int> durationMatrix_;
 
 int main() {
